+++ src/app_window.cpp
@@ -210,7 +210,7 @@ int AppWindow::run(HINSTANCE instance, int showCommand) {
   brandClass.lpszClassName = kBrandClass;
   RegisterClassExW(&brandClass);
 
-  window_ = CreateWindowExW(0, kMainClass, L"FUBAR VOX V1.1.2", WS_OVERLAPPEDWINDOW,
+  window_ = CreateWindowExW(0, kMainClass, L"FUBAR VOX V1.1.3", WS_OVERLAPPEDWINDOW,
                             CW_USEDEFAULT, CW_USEDEFAULT, 780, 780, nullptr, nullptr, instance_,
                             this);
   if (!window_) return 1;
@@ -543,6 +543,7 @@ void AppWindow::startEngine() {
     SetWindowTextW(outputEdit_, options_.outputDirectory.wstring().c_str());
   }
   web_.setRoot(options_.outputDirectory);
+  web_.setLiveHub(&engine_.liveHub());
   EnableWindow(startButton_, TRUE);
   EnableWindow(stopButton_, TRUE);
   updateStatus(L"Starting...");
@@ -774,6 +775,7 @@ void AppWindow::loadSettings() {
 }
 
 void AppWindow::applyWebServer() {
+  web_.setLiveHub(&engine_.liveHub());
   web_.setRoot(options_.outputDirectory);
   if (!webEnabled_) {
     web_.stop();